feat: embed substreams-sink-sql source under sink/sql/ (batch 1)
